Move around files to commands and minimize the vcf command call
