+++ commands/run/main.py
@@ -22,7 +22,7 @@ from configparser import ConfigParser
 from datetime import datetime
 from typing import Any, List, Dict, Optional
 
-from runner.gittools import (
+from commands.run.gittools import (
     check_if_on_branchhead,
     check_valid_checkout,
     check_valid_repo,
@@ -30,16 +30,16 @@ from runner.gittools import (
     get_git_commit_hash_and_log,
     pull_branch,
 )
-from util.shared_utils import check_valid_config_path, load_config
+from shared.util import check_valid_config_path, load_config
 
 from .help_classes import Case, CsvEntry
 
 
 logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
-LOG = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 
-ASSAY_PLACEHOLDER = "pipeeval"
+ASSAY_PLACEHOLDER = "dev"
 
 
 def main(
@@ -58,20 +58,18 @@ def main(
     datestamp: bool,
     verbose: bool,
 ):
-    LOG.info(f"Preparing run, type: {run_type}, data: {start_data}")
+    logger.info(f"Preparing run, type: {run_type}, data: {start_data}")
 
     check_valid_config_arguments(config, run_type, start_data, base_dir, repo)
-    base_dir = (
-        base_dir if base_dir is not None else Path(config.get("settings", "base"))
-    )
+    base_dir = base_dir if base_dir is not None else Path(config.get("settings", "base"))
     repo = repo if repo is not None else Path(config.get("settings", "repo"))
     datestamp = datestamp or config.getboolean("settings", "datestamp")
 
     check_valid_repo(repo)
-    check_valid_checkout(LOG, repo, checkout, verbose)
-    LOG.info(f"Checking out: {checkout} in {str(repo)}")
-    checkout_repo(LOG, repo, checkout, verbose)
-    on_branch_head = check_if_on_branchhead(LOG, repo, verbose)
+    check_valid_checkout(logger, repo, checkout, verbose)
+    logger.info(f"Checking out: {checkout} in {str(repo)}")
+    checkout_repo(logger, repo, checkout, verbose)
+    on_branch_head = check_if_on_branchhead(logger, repo, verbose)
     if on_branch_head:
         branch = checkout
         if skip_confirmation:
@@ -81,10 +79,10 @@ def main(
                 f"You have checked out the branch {branch} in {repo}. Do you want to pull? (y/n) "
             )
         if confirmation == "y":
-            LOG.info("Pulling from origin")
-            pull_branch(LOG, repo, branch, verbose)
-    (commit_hash, last_log) = get_git_commit_hash_and_log(LOG, repo, verbose)
-    LOG.info(last_log)
+            logger.info("Pulling from origin")
+            pull_branch(logger, repo, branch, verbose)
+    (commit_hash, last_log) = get_git_commit_hash_and_log(logger, repo, verbose)
+    logger.info(last_log)
 
     run_label = build_run_label(run_type, checkout, label, stub_run, start_data)
 
@@ -99,7 +97,7 @@ def main(
         )
 
         if confirmation != "y":
-            LOG.info("Exiting ...")
+            logger.info("Exiting ...")
             sys.exit(1)
 
     if not dry_run:
@@ -121,13 +119,9 @@ def main(
     run_type_settings = dict(config[run_type])
 
     if not config.getboolean(run_type, "trio"):
-        csv = get_single_csv(
-            config, run_type_settings, run_label, start_data, queue, stub_run
-        )
+        csv = get_single_csv(config, run_type_settings, run_label, start_data, queue, stub_run)
     else:
-        csv = get_trio_csv(
-            config, run_type_settings, run_label, start_data, queue, stub_run
-        )
+        csv = get_trio_csv(config, run_type_settings, run_label, start_data, queue, stub_run)
     out_csv = results_dir / "run.csv"
     if dry_run:
         logging.info(f"(dry) Writing CSV to {out_csv}")
@@ -205,9 +199,7 @@ def build_run_label(
     run_label = "-".join(label_parts)
 
     if run_label.find("/") != -1:
-        LOG.warning(
-            f"# Found '/' characters in run label: {run_label}, replacing with '-'"
-        )
+        logger.warning(f"Found '/' characters in run label: {run_label}, replacing with '-'")
         run_label = run_label.replace("/", "-")
 
     return run_label
@@ -276,9 +268,7 @@ def get_trio_csv(
 ):
 
     case_ids = run_type_settings["cases"].split(",")
-    assert (
-        len(case_ids) == 3
-    ), f"For a trio, three fields are expected, found: {case_ids}"
+    assert len(case_ids) == 3, f"For a trio, three fields are expected, found: {case_ids}"
     cases: List[Case] = []
     for case_id in case_ids:
         case_dict = config[case_id]
@@ -292,9 +282,7 @@ def get_trio_csv(
         case = parse_case(dict(case_dict), start_data, is_trio=True)
 
         if not Path(case.read1).exists() or not Path(case.read2).exists():
-            raise FileNotFoundError(
-                f"One or both files missing: {case.read1} {case.read2}"
-            )
+            raise FileNotFoundError(f"One or both files missing: {case.read1} {case.read2}")
 
         cases.append(case)
 
@@ -315,9 +303,7 @@ def parse_case(case_dict: Dict[str, str], start_data: str, is_trio: bool) -> Cas
         fw = case_dict["fq_fw"]
         rv = case_dict["fq_rv"]
     else:
-        raise ValueError(
-            f"Unknown start_data, found: {start_data}, valid are vcf, bam, fq"
-        )
+        raise ValueError(f"Unknown start_data, found: {start_data}, valid are vcf, bam, fq")
 
     case = Case(
         case_dict["id"],
@@ -385,9 +371,7 @@ def build_start_nextflow_analysis_cmd(
     return start_nextflow_command
 
 
-def start_run(
-    start_nextflow_command: List[str], dry_run: bool, skip_confirmation: bool
-):
+def start_run(start_nextflow_command: List[str], dry_run: bool, skip_confirmation: bool):
     if not dry_run:
         if not skip_confirmation:
             pretty_command = " \\\n    ".join(start_nextflow_command)
@@ -398,12 +382,12 @@ def start_run(
             if confirmation == "y":
                 subprocess.run(start_nextflow_command, check=True)
             else:
-                LOG.info("Exiting ...")
+                logger.info("Exiting ...")
         else:
             subprocess.run(start_nextflow_command, check=True)
     else:
         joined_command = " ".join(start_nextflow_command)
-        LOG.info("(dry) " + joined_command)
+        logger.info("(dry) " + joined_command)
 
 
 def write_resume_script(
@@ -424,9 +408,7 @@ def write_resume_script(
         logging.info(f"(dry) Writing {resume_command} to {resume_script}")
 
 
-def setup_results_links(
-    config: ConfigParser, results_dir: Path, run_label: str, dry: bool
-):
+def setup_results_links(config: ConfigParser, results_dir: Path, run_label: str, dry: bool):
 
     log_base_dir = config["settings"]["log_base_dir"]
     trace_base_dir = config["settings"]["trace_base_dir"]
@@ -439,26 +421,22 @@ def setup_results_links(
     trace_link = results_dir / "trace.txt"
     work_link = results_dir / "work"
 
-    log_link_target = Path(
-        f"{log_base_dir}/{run_label}.{ASSAY_PLACEHOLDER}.{date_stamp}.log"
-    )
-    trace_link_target = Path(
-        f"{trace_base_dir}/{run_label}.{ASSAY_PLACEHOLDER}.trace.txt"
-    )
+    log_link_target = Path(f"{log_base_dir}/{run_label}.{ASSAY_PLACEHOLDER}.{date_stamp}.log")
+    trace_link_target = Path(f"{trace_base_dir}/{run_label}.{ASSAY_PLACEHOLDER}.trace.txt")
     work_link_target = Path(f"{work_base_dir}/{run_label}.{ASSAY_PLACEHOLDER}")
 
     if log_link.exists():
-        LOG.warning(f"{log_link} already exists, removing previous link")
+        logger.warning(f"{log_link} already exists, removing previous link")
         if not dry:
             log_link.unlink()
 
     if trace_link.exists():
-        LOG.warning(f"{trace_link} already exists, removing previous link")
+        logger.warning(f"{trace_link} already exists, removing previous link")
         if not dry:
             trace_link.unlink()
 
     if work_link.exists():
-        LOG.warning(f"{work_link} already exists, removing previous link")
+        logger.warning(f"{work_link} already exists, removing previous link")
         if not dry:
             work_link.unlink()
 
@@ -467,20 +445,18 @@ def setup_results_links(
         trace_link.symlink_to(trace_link_target)
         work_link.symlink_to(work_link_target)
     else:
-        LOG.info(f"(dry) Linking log from {log_link_target} to {log_link}")
-        LOG.info(f"(dry) Linking trace from {trace_link_target} to {trace_link}")
-        LOG.info(f"(dry) Linking work from {work_link_target} to {work_link}")
+        logger.info(f"(dry) Linking log from {log_link_target} to {log_link}")
+        logger.info(f"(dry) Linking trace from {trace_link_target} to {trace_link}")
+        logger.info(f"(dry) Linking work from {work_link_target} to {work_link}")
 
 
 def main_wrapper(args: argparse.Namespace):
 
     curr_dir = os.path.dirname(os.path.abspath(__file__))
-    config = load_config(curr_dir, args.config)
+    config = load_config(logger, curr_dir, args.config)
 
     if args.baseline is not None:
-        logging.info(
-            "Performing additional baseline run as specified by --baseline flag"
-        )
+        logging.info("Performing additional baseline run as specified by --baseline flag")
 
         if args.baseline_repo is not None:
             baseline_repo = str(args.baseline_repo)
@@ -583,12 +559,8 @@ def add_arguments(parser: argparse.ArgumentParser):
         action="store_true",
         help="Run start_nextflow_analysis.pl with nostart, printing the path to the SLURM job only",
     )
-    parser.add_argument(
-        "--baseline", help="Start a second baseline run and specified checkout"
-    )
-    parser.add_argument(
-        "--verbose", action="store_true", help="Print additional debug output"
-    )
+    parser.add_argument("--baseline", help="Start a second baseline run and specified checkout")
+    parser.add_argument("--verbose", action="store_true", help="Print additional debug output")
 
 
 if __name__ == "__main__":
